renamed controller package to player_input_output
