+++ core/src/main/java/com/strategy_bit/chaos_brawl/player_input_output/PawnController.java
@@ -1,4 +1,4 @@
-package com.strategy_bit.chaos_brawl.controller;
+package com.strategy_bit.chaos_brawl.player_input_output;
 
 
 import com.strategy_bit.chaos_brawl.ResourceSystem.Resource;
