+++ core/src/main/java/com/strategy_bit/chaos_brawl/player_input_output/views/GameHUD.java
@@ -1,4 +1,4 @@
-package com.strategy_bit.chaos_brawl.views;
+package com.strategy_bit.chaos_brawl.player_input_output.views;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.InputProcessor;
